GUI: * Added support for selecting objects and editing its parameters in the level view tab.
Code:
* Changed some code relating to alpha handling in textures. (See bottom of log for more info)
* Fixed a crash involving 64 bit pointers being casted as an int.

Alpha support isn't there yet. Some things that should be fully transparent aren't. There is also an issue in the viewport where transparent textures will hide geometry behind them.

+++ trunk/WorldExplorer/WorldExplorer/DataLoaders/TexDecoder.cs
@@ -110,7 +110,7 @@ namespace WorldExplorer.DataLoaders
                     null);
                 image.Lock();
                 unsafe {
-                    int pBackBuffer = (int)image.BackBuffer;
+                    IntPtr pBackBuffer = image.BackBuffer;
                     for (int y = 0; y < sourceh; ++y) {
                         for (int x = 0; x < sourcew; ++x) {
                             PalEntry pixel = pixels[y * sourcew + x];
